[Universal] Changing some normal logic to use floats instead of halfs
The classic half to float PR. In this case normals were causing some minor artifacts that needed to be calculated as floats. These two changes are the bare minimum and should only affect half-supported platforms. 

Fix for UUM-11468

+++ Packages/com.unity.render-pipelines.universal/ShaderLibrary/Input.hlsl
@@ -33,7 +33,7 @@ struct InputData
 {
     float3  positionWS;
     float4  positionCS;
-    half3   normalWS;
+    float3  normalWS;
     half3   viewDirectionWS;
     float4  shadowCoord;
     half    fogCoord;
